able to train but serial

+++ operating_platform/policy/act/modeling_act.py
@@ -73,7 +73,8 @@ class ACTPolicy(PreTrainedPolicy):
         )
 
         self.model = ACT(config)
-        self.model = self.model.half()
+        # self.model = self.model.half()
+        self.model = self.model
 
         if config.temporal_ensemble_coeff is not None:
             self.temporal_ensembler = ACTTemporalEnsembler(config.temporal_ensemble_coeff, config.chunk_size)
@@ -342,12 +343,14 @@ class ACT(nn.Module):
                 weights=config.pretrained_backbone_weights,
                 norm_layer=FrozenBatchNorm2d,
             )
-            backbone_model = backbone_model.half()
+            # backbone_model = backbone_model.half()
+            backbone_model = backbone_model
             # Note: The assumption here is that we are using a ResNet model (and hence layer4 is the final
             # feature map).
             # Note: The forward method of this returns a dict: {"feature_map": output}.
             self.backbone = IntermediateLayerGetter(backbone_model, return_layers={"layer4": "feature_map"})
-            self.backbone = self.backbone.half()
+            # self.backbone = self.backbone.half()
+            self.backbone = self.backbone
         # Transformer (acts as VAE decoder when training with the variational objective).
         self.encoder = ACTEncoder(config)
         self.decoder = ACTDecoder(config)
@@ -492,7 +495,8 @@ class ACT(nn.Module):
             # NOTE: If modifying this section, verify on MPS devices that
             # gradients remain stable (no explosions or NaNs).
             for img in batch["observation.images"]:
-                cam_features = self.backbone(img.half())["feature_map"]
+                # cam_features = self.backbone(img.half())["feature_map"]
+                cam_features = self.backbone(img)["feature_map"]
                 cam_pos_embed = self.encoder_cam_feat_pos_embed(cam_features).to(dtype=cam_features.dtype)
                 cam_features = self.encoder_img_feat_input_proj(cam_features)
