+++ operating_platform/policy/pretrained.py
@@ -125,7 +125,8 @@ class PreTrainedPolicy(nn.Module, HubMixin, abc.ABC):
 
         policy.to(config.device)
         policy.eval()
-        policy = policy.half()
+        # policy = policy.half()
+        policy=policy
         return policy
 
     @classmethod
